fix duplicate _key detected, track duplicate keys count, refactor code
