+++ src/steps/containers/index.ts
@@ -1,259 +1,15 @@
 import {
-  createDirectRelationship,
-  Entity,
-  getRawData,
   IntegrationStepExecutionContext,
-  RelationshipClass,
   Step,
 } from '@jupiterone/integration-sdk-core';
 import { IntegrationConfig } from '../../config';
+import { Entities, Relationships, StepIds } from '../constants';
 import {
-  Entities,
-  Relationships,
-  SERVICE_ENTITY_DATA_KEY,
-  StepIds,
-} from '../constants';
-import {
-  containerFindingEntityKey,
-  createReportUnwantedProgramRelationship,
-  createMalwareEntity,
-  createReportMalwareRelationship,
-  createUnwantedProgramEntity,
-  malwareEntityKey,
-  unwantedProgramEntityKey,
-  createContainerRepositoryEntity,
-  createAccountContainerRepositoryRelationship,
-} from './converters';
-import { getAccount } from '../account/util';
-import TenableClient from '../../tenable/TenableClient';
-import { ContainerImage } from '../../tenable/client';
-import {
-  createAccountContainerImageRelationship,
-  createContainerImageEntity,
-  createContainerFindingEntity,
-  createContainerReportRelationship,
-  createReportEntity,
-  createReportFindingRelationship,
-} from './converters';
-import { generateEntityKey } from '../../utils/generateKey';
-
-export async function fetchContainerRepositories(
-  context: IntegrationStepExecutionContext<IntegrationConfig>,
-): Promise<void> {
-  const {
-    jobState,
-    logger,
-    instance: {
-      config: { accessKey, secretKey },
-    },
-  } = context;
-  const client = new TenableClient({
-    logger,
-    accessToken: accessKey,
-    secretToken: secretKey,
-  });
-
-  const account = getAccount(context);
-  await client.iterateContainerRepositories(async (repository) => {
-    await jobState.addEntity(createContainerRepositoryEntity(repository));
-    await jobState.addRelationship(
-      createAccountContainerRepositoryRelationship(account, repository),
-    );
-  });
-}
-
-export async function fetchContainerImages(
-  context: IntegrationStepExecutionContext<IntegrationConfig>,
-): Promise<void> {
-  const { jobState, logger, instance } = context;
-  const client = new TenableClient({
-    logger: logger,
-    accessToken: instance.config.accessKey,
-    secretToken: instance.config.secretKey,
-  });
-
-  const serviceEntity = (await jobState.getData(
-    SERVICE_ENTITY_DATA_KEY,
-  )) as Entity;
-
-  const account = getAccount(context);
-  await client.iterateContainerImages(async (image) => {
-    const imageEntity = createContainerImageEntity(image);
-    await Promise.all([
-      jobState.addEntity(imageEntity),
-      jobState.addRelationship(
-        createAccountContainerImageRelationship(account, image),
-      ),
-      jobState.addRelationship(
-        createDirectRelationship({
-          _class: RelationshipClass.SCANS,
-          from: serviceEntity,
-          to: imageEntity,
-        }),
-      ),
-    ]);
-  });
-}
-
-export async function buildRepositoryImagesRelationship(
-  context: IntegrationStepExecutionContext<IntegrationConfig>,
-): Promise<void> {
-  const { jobState, logger } = context;
-
-  await jobState.iterateEntities(
-    { _type: Entities.CONTAINER_IMAGE._type },
-    async (imageEntity) => {
-      const image = getRawData<ContainerImage>(imageEntity);
-
-      if (!image) {
-        logger.warn(
-          {
-            _key: imageEntity._key,
-          },
-          'Could not fetch raw data for container image entity',
-        );
-        return;
-      }
-      const { repoName } = image;
-      const repoEntity = await jobState.findEntity(
-        generateEntityKey(Entities.CONTAINER_REPOSITORY._type, repoName),
-      );
-      if (repoEntity) {
-        await jobState.addRelationship(
-          createDirectRelationship({
-            _class: RelationshipClass.HAS,
-            from: repoEntity,
-            to: imageEntity,
-          }),
-        );
-      }
-    },
-  );
-}
-
-export async function fetchContainerReports(
-  context: IntegrationStepExecutionContext<IntegrationConfig>,
-): Promise<void> {
-  const { jobState, logger, instance } = context;
-  const client = new TenableClient({
-    logger: logger,
-    accessToken: instance.config.accessKey,
-    secretToken: instance.config.secretKey,
-  });
-
-  await jobState.iterateEntities(
-    { _type: Entities.CONTAINER_IMAGE._type },
-    async (imageEntity) => {
-      const image = getRawData<ContainerImage>(imageEntity);
-
-      if (!image) {
-        logger.warn(
-          {
-            _key: imageEntity._key,
-          },
-          'Could not fetch raw data for container image entity',
-        );
-        return;
-      }
-      const { repoName, name: imageName, tag } = image;
-      const report = await client.fetchContainerImageReport(
-        repoName,
-        imageName,
-        tag,
-      );
-      await jobState.addEntity(createReportEntity(report));
-      await jobState.addRelationship(
-        createContainerReportRelationship(image, report),
-      );
-
-      for (const finding of report.findings) {
-        const findingKey = containerFindingEntityKey(finding);
-        let findingEntity = await jobState.findEntity(findingKey);
-
-        if (!findingEntity) {
-          findingEntity = await jobState.addEntity(
-            createContainerFindingEntity(finding),
-          );
-        }
-
-        const reportFindingRelationship = createReportFindingRelationship(
-          report.sha256,
-          finding,
-        );
-
-        if (!jobState.hasKey(reportFindingRelationship._key)) {
-          await jobState.addRelationship(reportFindingRelationship);
-        }
-
-        const imageFindingRelationship = createDirectRelationship({
-          from: imageEntity,
-          _class: RelationshipClass.HAS,
-          to: findingEntity,
-        });
-
-        if (!jobState.hasKey(imageFindingRelationship._key)) {
-          await jobState.addRelationship(imageFindingRelationship);
-        }
-      }
-
-      for (const malware of report.malware) {
-        const malwareKey = malwareEntityKey(malware);
-        let malwareEntity = await jobState.findEntity(malwareKey);
-
-        if (!malwareEntity) {
-          malwareEntity = await jobState.addEntity(
-            createMalwareEntity(malware),
-          );
-        }
-
-        const reportMalwareRelationship = createReportMalwareRelationship(
-          report.sha256,
-          malware,
-        );
-        if (!jobState.hasKey(reportMalwareRelationship._key)) {
-          await jobState.addRelationship(reportMalwareRelationship);
-        }
-
-        const imageMalwareRelationship = createDirectRelationship({
-          from: imageEntity,
-          _class: RelationshipClass.HAS,
-          to: malwareEntity,
-        });
-
-        if (!jobState.hasKey(imageMalwareRelationship._key))
-          await jobState.addRelationship(imageMalwareRelationship);
-      }
-
-      for (const program of report.potentially_unwanted_programs) {
-        const programKey = unwantedProgramEntityKey(program);
-        let programEntity = await jobState.findEntity(programKey);
-
-        if (!programEntity) {
-          programEntity = await jobState.addEntity(
-            createUnwantedProgramEntity(program),
-          );
-        }
-
-        const reportUnwantedProgramRelationship =
-          createReportUnwantedProgramRelationship(report.sha256, program);
-
-        if (!jobState.hasKey(reportUnwantedProgramRelationship._key)) {
-          await jobState.addRelationship(reportUnwantedProgramRelationship);
-        }
-
-        const imageProgramRelationship = createDirectRelationship({
-          from: imageEntity,
-          _class: RelationshipClass.HAS,
-          to: programEntity,
-        });
-
-        if (!jobState.hasKey(imageProgramRelationship._key)) {
-          await jobState.addRelationship(imageProgramRelationship);
-        }
-      }
-    },
-  );
-}
+  buildRepositoryImagesRelationship,
+  fetchContainerImages,
+  fetchContainerReports,
+  fetchContainerRepositories,
+} from './handlers';
 
 export const containerSteps: Step<
   IntegrationStepExecutionContext<IntegrationConfig>
